Move about page tools/jobs section to components and move logos.
