+++ gatsby-config.js
@@ -2,7 +2,7 @@ module.exports = {
   siteMetadata: {
     title: 'Brad McGonigle',
     author: 'Brad McGonigle',
-    description: 'Developer, nerd and husband.',
+    description: 'Father, husband, developer, and nerd.',
     siteUrl: 'https://www.bradmcgonigle.com',
   },
   plugins: [
