+++ src/components/about/career/tools/logos/tools-logo-builder.sketch
